Split Rule and Shape to separate modules
