+++ matrices/types.pyi
@@ -1,17 +1,14 @@
-from collections.abc import Collection, Iterable, Sequence
-from enum import IntEnum
+from collections.abc import Iterable, Sequence
 from typing import (Any, Iterator, Literal, Optional, SupportsIndex, TypeVar,
-                    Union, final, overload)
+                    Union, overload)
 
 from .protocols import (ComplexLike, ComplexMatrixLike, IntegralLike,
                         IntegralMatrixLike, MatrixLike, RealLike,
-                        RealMatrixLike, ShapeLike)
+                        RealMatrixLike)
+from .rule import Rule
+from .shape import Shape
 
 __all__ = [
-    "Rule",
-    "ROW",
-    "COL",
-    "Shape",
     "GenericMatrix",
     "ComplexMatrix",
     "RealMatrix",
@@ -24,63 +21,6 @@ RealLikeT = TypeVar("RealLikeT", bound=RealLike)
 IntegralLikeT = TypeVar("IntegralLikeT", bound=IntegralLike)
 
 Self = TypeVar("Self")
-
-
-@final
-class Rule(IntEnum):
-
-    ROW: int
-    COL: int
-
-    @property
-    def inverse(self: Self) -> Rule: ...
-    @property
-    def true_name(self: Self) -> Literal["row", "column"]: ...
-
-
-ROW: Literal[Rule.ROW]
-COL: Literal[Rule.COL]
-
-
-@final
-class Shape(Collection[int]):
-
-    __slots__: tuple[Literal["data"]]
-
-    def __init__(self: Self, nrows: int = 0, ncols: int = 0) -> None: ...
-    def __repr__(self: Self) -> str: ...
-    def __str__(self: Self) -> str: ...
-    def __eq__(self: Self, other: Any) -> bool: ...
-    def __ne__(self: Self, other: Any) -> bool: ...
-    def __len__(self: Self) -> Literal[2]: ...
-    def __getitem__(self: Self, key: SupportsIndex) -> int: ...
-    def __setitem__(self: Self, key: SupportsIndex, value: int) -> None: ...
-    def __iter__(self: Self) -> Iterator[int]: ...
-    def __reversed__(self: Self) -> Iterator[int]: ...
-    def __contains__(self: Self, value: Any) -> bool: ...
-    def __deepcopy__(self: Self, memo: Optional[dict[int, Any]] = None) -> Shape: ...
-    def __copy__(self: Self) -> Shape: ...
-
-    @property
-    def nrows(self: Self) -> int: ...
-    @nrows.setter
-    def nrows(self: Self, value: int) -> None: ...
-    @property
-    def ncols(self: Self) -> int: ...
-    @ncols.setter
-    def ncols(self: Self, value: int) -> None: ...
-    @property
-    def size(self: Self) -> int: ...
-
-    def true_equals(self: Self, other: ShapeLike) -> bool: ...
-    def copy(self: Self) -> Shape: ...
-    def reverse(self: Self) -> Shape: ...
-    def subshape(self: Self, *, by: Rule = Rule.ROW) -> Shape: ...
-    def resolve_index(self: Self, key: SupportsIndex, *, by: Rule = Rule.ROW) -> int: ...
-    def resolve_slice(self: Self, key: slice, *, by: Rule = Rule.ROW) -> range: ...
-    def sequence(self: Self, index: int, *, by: Rule = Rule.ROW) -> tuple[int, int, int]: ...
-    def range(self: Self, index: int, *, by: Rule = Rule.ROW) -> range: ...
-    def slice(self: Self, index: int, *, by: Rule = Rule.ROW) -> slice: ...
 
 
 class GenericMatrix(Sequence[T]):
